Rename from PascalCase to kebab-case
