+++ src/js/object-descriptor.js
